requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2==2.9.9
drf-spectacular==0.27
Pillow==10.2.0